Changed internal working of extension system.
Most fields and methods of the extensible classes are now put in
a class with the 'Basic' prefix. The extensions inherit that, and the
final class inherits the last extended one.

This way, the extensions can access the fields and methods that are
in the Basic one. Giving the 'extender' a little bit more freedom.